ref(node): move Config to a subpackage
